rename WithShippingRewardsCollectionViewHeaderView
